Rename all V1 files
