+++ .github/workflows/vllm_ascend_test.yaml
@@ -203,12 +203,12 @@ jobs:
           --ignore=tests/e2e/singlecard/test_guided_decoding.py \
           --ignore=tests/e2e/singlecard/test_camem.py \
           --ignore=tests/e2e/singlecard/test_embedding.py \
-          --ignore=tests/e2e/singlecard/spec_decode_v1/test_v1_mtp_correctness.py \
-          --ignore=tests/e2e/singlecard/spec_decode_v1/test_v1_spec_decode.py
-          # ------------------------------------ v1 spec decode test ------------------------------------ #
-          VLLM_USE_MODELSCOPE=True pytest -sv tests/e2e/singlecard/spec_decode_v1/test_v1_mtp_correctness.py
-          # TODO: revert me when test_v1_spec_decode.py::test_ngram_correctness is fixed
-          VLLM_USE_MODELSCOPE=True pytest -sv tests/e2e/singlecard/spec_decode_v1/test_v1_spec_decode.py
+          --ignore=tests/e2e/singlecard/spec_decode/test_mtp_correctness.py \
+          --ignore=tests/e2e/singlecard/spec_decode/test_spec_decode.py
+          # ------------------------------------ spec decode test ------------------------------------ #
+          VLLM_USE_MODELSCOPE=True pytest -sv tests/e2e/singlecard/spec_decode/test_mtp_correctness.py
+          # TODO: revert me when test_spec_decode.py::test_ngram_correctness is fixed
+          VLLM_USE_MODELSCOPE=True pytest -sv tests/e2e/singlecard/spec_decode/test_spec_decode.py
 
   e2e-4-cards:
     needs: [e2e]
