+++ vllm_ascend/quantization/w8a8.py
@@ -22,7 +22,7 @@ import torch_npu
 from vllm.attention.backends.abstract import AttentionType
 from vllm.distributed.parallel_state import get_ep_group
 
-from vllm_ascend.attention.attention_v1 import AscendAttentionState
+from vllm_ascend.attention.attention import AscendAttentionState
 from vllm_ascend.utils import ACL_FORMAT_FRACTAL_NZ, is_310p
