+++ vllm_ascend/worker/eagle_proposer.py
@@ -14,8 +14,8 @@ from vllm.model_executor.models import supports_multimodal
 from vllm.model_executor.models.llama_eagle3 import Eagle3LlamaForCausalLM
 from vllm.v1.sample.metadata import SamplingMetadata
 
+from vllm_ascend.attention.attention import AscendAttentionState
 from vllm_ascend.attention.attention_mask import AttentionMaskBuilder
-from vllm_ascend.attention.attention_v1 import AscendAttentionState
 
 logger = init_logger(__name__)
